Rearrange tests
Refs: https://github.com/metarhia/metasync/issues/258
