+++ lib/control.js
@@ -43,7 +43,6 @@ const parallel = (
     if (result !== context && result !== undefined) {
       if (isArray) context.push(result);
       else if (typeof(result) === 'object') Object.assign(context, result);
-      else if (fn.name) context[fn.name] = result;
     }
     if (++counter === len) done(null, context);
   };
@@ -82,7 +81,6 @@ const sequential = (
       if (result !== context && result !== undefined) {
         if (isArray) context.push(result);
         else if (typeof(result) === 'object') Object.assign(context, result);
-        else if (fn.name) context[fn.name] = result;
       }
       if (err) {
         done(err);
