+++ lib/flow.js
@@ -126,7 +126,6 @@ function parallel(
     if (result !== context && result !== undefined) {
       if (isArray) context.push(result);
       else if (typeof(result) === 'object') Object.assign(context, result);
-      //else if (fn.name) context[fn.name] = result;
     }
     if (++counter === len) done(null, context);
   };
@@ -166,7 +165,6 @@ function sequential(
       if (result !== context && result !== undefined) {
         if (isArray) context.push(result);
         else if (typeof(result) === 'object') Object.assign(context, result);
-        //else if (fn.name) context[fn.name] = result;
       }
       if (err) {
         done(err);
