+++ test/flow.js
@@ -27,52 +27,6 @@ tap.test('flow with parallel flow', (test) => {
   });
 });
 
-tap.test('parallel with error', (test) => {
-  const parallelError = new Error('Parallel error');
-
-  function fn1(data, cb) {
-    process.nextTick(() => {
-      cb(null, { data1: 'data 1' });
-    });
-  }
-
-  function fn2(data, cb) {
-    process.nextTick(() => {
-      cb(parallelError);
-    });
-  }
-
-  metasync.parallel([fn1, fn2], (err, res) => {
-    test.strictSame(err, parallelError);
-    test.strictSame(res, undefined);
-    test.end();
-  });
-});
-
-tap.test('sequential with error', (test) => {
-  const sequentialError = new Error('Sequential error');
-  const expectedDataInFn2 = { data1: 'data 1' };
-
-  function fn1(data, cb) {
-    process.nextTick(() => {
-      cb(null, { data1: 'data 1' });
-    });
-  }
-
-  function fn2(data, cb) {
-    process.nextTick(() => {
-      tap.same(data, expectedDataInFn2);
-      cb(sequentialError);
-    });
-  }
-
-  metasync.sequential([fn1, fn2], (err, res) => {
-    test.strictSame(err, sequentialError);
-    test.strictSame(res, undefined);
-    test.end();
-  });
-});
-
 tap.test('flow with complex flow', (test) => {
 
   const data = { test: 'data' };
